Refactorización de microservicio para definir RUN como ID
Se realiza refactorización para convertir RUN de usuario como ID

+++ usuarios-service/src/main/java/petly/sanosysalvos/cl/usuarios/Repository/usuarioRepository.java
@@ -6,10 +6,12 @@ import  org.springframework.data.jpa.repository.JpaRepository;
 
 import petly.sanosysalvos.cl.usuarios.Model.Usuario;
 
-public interface usuarioRepository extends JpaRepository<Usuario, Long>{
+public interface usuarioRepository extends JpaRepository<Usuario, Integer>{
 
     // Método personalizado para buscar un usuario por correo
     Optional<Usuario> findByCorreo(String correo);
+
+
 
 
 }
